+++ inkscape_qrcode/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # Inkscape QR Code
-# Copyright (C) 2016 - 2017 Lars Heuer
+# Copyright (C) 2016 - 2018 Lars Heuer
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of version 2 of the GNU General Public