+++ inkscape_qrcode/_segno/__init__.py
@@ -6,4 +6,4 @@
 # License: BSD License
 #
 
-__version__ = '0.2.4'
+__version__ = '0.2.7'